Share the test clock and take the status a caller already read
The MutableClock in the two lease tests was the same class twice. Moving
it out of MongoLeaseRaceTest and MongoLeaseTimingTest leaves one copy for
both.

acquireLease read the consumer's status again although every caller had
just read it under the same lock, so it takes that status instead, which
also puts it and giveUpLease on the same footing.
